sessionID implement with local storage for Chatbot component
